Concert Report Page implementation started
